Rafactor, catch mongo errros

+++ test/receiver.test.ts
@@ -3,30 +3,44 @@ import * as receiver from '../lib/receiver'
 import * as channel from '../lib/channel'
 import { randomStorage } from './support'
 import { PaymentChannel } from '../lib/channel'
-import mongo from '../lib/mongo';
+import mongo from '../lib/mongo'
+
+const engine_name = process.env.ENGINE_NAME || 'nedb'
 
 describe('receiver', () => {
   beforeAll((done) => {
-    mongo.connectToServer( () => {
+    if (process.env.ENGINE_NAME == 'mongo') {
+      mongo.connectToServer(() => {
+        done()
+      })
+    } else {
       done()
-    });
-  });
+    }
+  })
 
   beforeEach((done) => {
-    mongo.db().dropDatabase(() => {
+    if (process.env.ENGINE_NAME == 'mongo') {
+      mongo.db().dropDatabase(() => {
+        done()
+      })
+    } else {
       done()
-    })
-  });
+    }
+  })
 
   afterAll((done) => {
-    mongo.db().close()
-  });
+    if (process.env.ENGINE_NAME == 'mongo') {
+      mongo.db().close()
+    } else {
+      done()
+    }
+  })
 
   let web3 = support.fakeWeb3()
 
   describe('.build', () => {
     it('build Receiver', done => {
-      randomStorage(web3).then(storage => {
+      randomStorage(web3, engine_name).then(storage => {
         let result = receiver.build(support.fakeWeb3(), '0xdeadbeaf', storage)
         expect(typeof result).toBe('object')
       }).then(done)
@@ -51,7 +65,7 @@ describe('receiver', () => {
           s: '0x3'
         })
         let paymentChannel = channel.PaymentChannel.fromPayment(payment)
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           return storage.channels.save(paymentChannel).then(() => {
             return receiver.build(web3, '0xdeadbeaf', storage).findPaymentChannel(payment)
           }).then((found: PaymentChannel|null) => {
@@ -79,7 +93,7 @@ describe('receiver', () => {
           r: '0x2',
           s: '0x3'
         })
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           return receiver.build(web3, '0xdeadbeaf', storage).findPaymentChannel(payment).then(found => {
             expect(found).toBeNull()
           })
@@ -100,7 +114,7 @@ describe('receiver', () => {
         s: '0x3'
       })
       it('return token', done => {
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           return receiver.build(web3, '0xdeadbeaf', storage).whenValidPayment(payment).then(token => {
             expect(token).not.toBeNull()
           })
@@ -108,7 +122,7 @@ describe('receiver', () => {
       })
 
       it('save payment', () => {
-        return randomStorage(web3).then(storage => {
+        return randomStorage(web3, engine_name).then(storage => {
           return receiver.build(web3, '0xdeadbeaf', storage).whenValidPayment(payment).then(() => {
             return storage.payments.firstMaximum(payment.channelId)
           }).then(savedPayment => {
@@ -130,7 +144,7 @@ describe('receiver', () => {
       })
 
       it('save token', done => {
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           return receiver.build(web3, '0xdeadbeaf', storage).whenValidPayment(payment).then(token => {
             return storage.tokens.isPresent(token)
           }).then(isPresent => {
@@ -155,7 +169,7 @@ describe('receiver', () => {
       })
 
       it('check if token is present', done => {
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           let r = receiver.build(web3, '0xdeadbeaf', storage)
           return r.whenValidPayment(payment).then(token => {
             return r.acceptToken(token)
@@ -167,7 +181,7 @@ describe('receiver', () => {
 
       it('check if token is absent', done => {
         let randomToken = support.randomInteger().toString()
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           let r = receiver.build(web3, '0xdeadbeaf', storage)
           return r.acceptToken(randomToken).then(isPresent => {
             expect(isPresent).toBeFalsy()
@@ -191,7 +205,7 @@ describe('receiver', () => {
       })
 
       it('throw an error if can not', () => {
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           let r = receiver.build(web3, '0xdeadbeaf', storage)
           expect(() => {
             r.ensureCanAcceptPayment(payment)
@@ -216,7 +230,7 @@ describe('receiver', () => {
       })
 
       it('accept new payment, and return a token', done => {
-        randomStorage(web3).then(storage => {
+        randomStorage(web3, engine_name).then(storage => {
           let r = receiver.build(web3, receiverAccount, storage)
           return r.acceptPayment(payment).then(token => {
             expect(token).not.toBeNull()
